requirements.txt
Django==2.1.3
djangorestframework==3.9.0
beautifulsoup4==4.6.3